+++ libs/core/src/index.ts
@@ -28,6 +28,7 @@ export * from './lib/actions/scrapers'
 export * from './lib/actions/loops'
 export * from './lib/actions/branching'
 export * from './lib/actions/time'
+export * from './lib/actions/random'
 
 //
 // Helpers
@@ -36,8 +37,10 @@ export * from './lib/helpers/cases'
 export * from './lib/helpers/console'
 export * from './lib/helpers/files'
 export * from './lib/helpers/indexed-db'
+export * from './lib/helpers/input'
 export * from './lib/helpers/local-storage'
 export * from './lib/helpers/navigation'
 export * from './lib/helpers/pipe'
 export * from './lib/helpers/scrapers'
 export * from './lib/helpers/cases'
+export * from './lib/helpers/random'