+++ libs/core/src/lib/actions/indexed-db.ts
@@ -1,12 +1,25 @@
 
 import { BotAction, BotIndexedDBAction } from '../interfaces/bot-actions'
 import { unpipeInjects } from '../helpers/pipe'
-import { getIndexedDBStoreValue, setIndexedDBStoreValue } from '../helpers/indexed-db'
+import { getIndexedDBStoreValue, setIndexedDBStoreValue, deleteIndexedDBDatabase } from '../helpers/indexed-db'
 import { inject } from './inject'
 import { PipeValue } from '../types/pipe-value'
 import { isObjectWithKey, isObjectWithValue } from '../types/objects'
 import { getQueryKey, getQueryKeyValue } from '../types/database'
 import { pipe } from './assembly-lines'
+
+/**
+ *
+ * @param page
+ */
+export const deleteIndexedDB = (databaseName?: string): BotIndexedDBAction<void> => async(page, ...injects) => {
+  const [, , injectDatabaseName] = unpipeInjects(injects, 2)
+
+  await page.evaluate(
+    deleteIndexedDBDatabase,
+    databaseName ?? injectDatabaseName
+  )
+}
 
 /**
  * @description    It's a higher-order BotAction that sets injects for identifying information of one IndexedDB store
@@ -38,31 +51,18 @@ export const indexedDBStore = (databaseName: string, storeName: string, database
 export const setIndexedDBValue =
   (key?: string, value?: any, storeName?: string, databaseName?: string, databaseVersion?: number): BotIndexedDBAction<void> =>
     async(page, ...injects) => {
-      const [pipedValue, injectDatabaseVersion, injectDatabaseName, injectStoreName] = unpipeInjects<getQueryKeyValue>(injects, 3)
+      let [pipeValue, injectDatabaseVersion, injectDatabaseName, injectStoreName] = unpipeInjects<getQueryKeyValue>(injects, 3)
 
-      if (!value) {
-        if (pipedValue) {
-          // idea here is that the piped value is another object with keys {key: '', value: ''} -> to map as what we are setting in the DB
-          if (isObjectWithValue(pipedValue)) {
-            value = pipedValue.value
-          } else {
-            value = pipedValue
-          }
-        }
-      }
-      if (!key) {
-        if (isObjectWithKey(pipedValue)) {
-          key = pipedValue.key
-        }
-      }
+      value ??= isObjectWithValue(pipeValue) ? pipeValue.value : pipeValue ??= 'missing-value'
+      key ??= isObjectWithKey(pipeValue) ? pipeValue.key : 'missing-key'
 
       await page.evaluate(
         setIndexedDBStoreValue,
         databaseName ? databaseName : injectDatabaseName ?? 'missing-db-name',
         databaseVersion ? databaseVersion : injectDatabaseVersion ?? undefined, // grab latest version
         storeName ? storeName : injectStoreName ?? 'missing-store',
-        key ?? 'missing-key',
-        value ?? 'missing-value'
+        key,
+        value
       )
     }
 
@@ -78,17 +78,9 @@ export const setIndexedDBValue =
 export const getIndexedDBValue =
   (key?: string, storeName?: string, databaseName?: string, databaseVersion?: number): BotIndexedDBAction<PipeValue> =>
     async(page, ...injects) => {
-      const [pipeValue, injectDatabaseVersion, injectDatabaseName, injectStoreName] = unpipeInjects<getQueryKey>(injects, 3)
+      let [pipeValue, injectDatabaseVersion, injectDatabaseName, injectStoreName] = unpipeInjects<getQueryKey|string>(injects, 3)
 
-      if (!key) {
-        if (pipeValue) {
-          if (isObjectWithKey(pipeValue)) {
-            key = pipeValue.key
-          } else {
-            key = pipeValue
-          }
-        }
-      }
+      key ??= isObjectWithKey(pipeValue) ? pipeValue.key : pipeValue as string
 
       return page.evaluate(
         getIndexedDBStoreValue,