+++ libs/core/src/lib/types/objects.ts
@@ -1,12 +1,12 @@
 /**
- * Generic Object Typing to Help 
+ * Generic Object Typing to Help
  *    Dictionaries, key/value pairs, or just a key
  */
 
 /**
  * @description    Is data provided an Object with a `key` property that holds a `string` value
  *                 ie getting data in a DB
- * @param data 
+ * @param data
  */
 export const isObjectWithKey = (data: any): data is {key: string} => {
   if (!data) {
@@ -18,7 +18,7 @@ export const isObjectWithKey = (data: any): data is {key: string} => {
 
 /**
  * @description   Is data provided an Object with a `value` property
- * @param data 
+ * @param data
  */
 export const isObjectWithValue = (data: any): data is {value: any} => {
   if (!data) {
@@ -32,14 +32,14 @@ export const isObjectWithValue = (data: any): data is {value: any} => {
  * Basic Hash-Map type
  */
 export type Dictionary<V = any> = {
-  [key: string]: V 
+  [key: string]: V
 }
 
 /**
  * Dictionaries are non-null objects with at least one key/value
- * @param value 
+ * @param value
  */
-export const isDictionary = <V = any>(value: any): value is Dictionary<V> => 
+export const isDictionary = <V = any>(value: any): value is Dictionary<V> =>
   typeof value === 'object' && value !== null && !Array.isArray(value) && Object.keys(value).every(key => typeof key === 'string')
 
 /**