+++ libs/instagram/src/index.ts
@@ -1,9 +1,8 @@
 export * from './lib/actions/auth'
 export * from './lib/actions/modals'
+export * from './lib/actions/navigation'
+export * from './lib/actions/stories'
 
 export * from './lib/constants/modals'
 export * from './lib/constants/urls'
-
-export * from './lib/helpers/urls'
-
-export * from './lib/selectors'
+export * from './lib/constants/selectors'