+++ libs/instagram/src/lib/actions/auth.ts
@@ -1,20 +1,20 @@
-import { ConditionalBotAction } from '@botmation/core'
+import { clickText, ConditionalBotAction, pipe, textExists } from '@botmation/core'
 import { BotAction } from '@botmation/core'
 
-import { chain } from '@botmation/core'
-import { goTo, waitForNavigation, wait } from '@botmation/core'
+import { chain, errors } from '@botmation/core'
+import { goTo, reload, waitForNavigation, wait } from '@botmation/core'
+import { getCookies, deleteCookies } from '@botmation/core'
 import { click, type } from '@botmation/core'
-import { getIndexedDBValue, indexedDBStore } from '@botmation/core'
+import { getIndexedDBValue, indexedDBStore, clearAllLocalStorage } from '@botmation/core'
 import { map } from '@botmation/core'
 import { log } from '@botmation/core'
 
-import { getInstagramLoginUrl } from '../helpers/urls'
+import { INSTAGRAM_URL_LOGIN } from '../constants/urls'
 import {
   FORM_AUTH_USERNAME_INPUT_SELECTOR,
   FORM_AUTH_PASSWORD_INPUT_SELECTOR,
   FORM_AUTH_SUBMIT_BUTTON_SELECTOR
-} from '../selectors'
-import { errors } from '@botmation/core'
+} from '../constants/selectors'
 
 /**
  * @description    ConditionalBotAction that resolves TRUE if the User is NOT logged in
@@ -47,7 +47,7 @@ export const isLoggedIn: ConditionalBotAction =
 export const login = ({username, password}: {username: string, password: string}): BotAction =>
   chain(
     errors('Instagram login()')(
-      goTo(getInstagramLoginUrl()),
+      goTo(INSTAGRAM_URL_LOGIN),
       click(FORM_AUTH_USERNAME_INPUT_SELECTOR),
       type(username),
       click(FORM_AUTH_PASSWORD_INPUT_SELECTOR),
@@ -58,3 +58,30 @@ export const login = ({username, password}: {username: string, password: string}
       log('Login Complete')
     )
   )
+
+/**
+ *
+ * @param page
+ */
+export const logout: BotAction = pipe()(
+  getCookies(),
+  deleteCookies(),
+  clearAllLocalStorage,
+  reload() // instagram will delete IndexedDB 'redux' db after reload
+)
+
+/**
+ * During initial login, in a brand new environment, Instagram might prompt the User to save their login information
+ *  This will disrupt feed access, therefore this function can be used to check if that is being presented
+ */
+export const isSaveYourLoginInfoActive: ConditionalBotAction = textExists('Save Your Login Info?')
+
+/**
+ *
+ */
+export const clickSaveYourLoginInfoYesButton: BotAction = clickText('Save Info')
+
+/**
+ *
+ */
+export const clickSaveYourLoginInfoNoButton: BotAction = clickText('Not Now')